refactor: 실행 로직 입력값 객체에서 분리
