+++ src/test/java/domain/CarFactoryTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import util.CarFactory;
 
 import java.util.List;
